testing new exists function

+++ manifests/install.pp
@@ -99,29 +99,30 @@ class jasperreports_server::install (
       before       => File['default_master.properties'],
     }
   }
-  file { 'default_master.properties':
-    ensure  => present,
-    path    => "/tmp/jasperreports-server-cp-${pkg_version}-bin/buildomatic/default_master.properties",
-    owner   => $buildomatic_user,
-    group   => $buildomatic_user,
-    content => template('jasperreports_server/default_master.properties.erb'),
-    require => [ Archive::Nexus["/tmp/jasperreports-server-cp-${pkg_version}-bin.zip"], ],
-  } ->
-  # Run the js-install with minimal flag
-  exec { 'Run js-install minimal':
-    path    => "/bin:/usr/bin:/sbin:/usr/sbin:/tmp/jasperreports-server-cp-${pkg_version}-bin/buildomatic",
-    cwd     => "/tmp/jasperreports-server-cp-${pkg_version}-bin/buildomatic",
-    command => 'js-install-ce.sh minimal',
-    creates => "${buildomatic_appserverdir}/webapps/jasperserver",
-    user    => $buildomatic_user,
-  } ->
-  # Dirty hack because of issues getting js-install to run as non-root user
-  exec { 'Update Jasper WebApp Ownership':
-    path        => '/bin:/usr/bin:/sbin:/usr/sbin',
-    command     => "chown -R tomcat:tomcat ${buildomatic_appserverdir}/webapps/jasperserver",
-    unless      => "stat -c %U ${buildomatic_appserverdir}/webapps/jasperserver | grep -q 'tomcat'",
-    subscribe   => Exec['Run js-install minimal'],
-    refreshonly => true,
+  if ! exists(${buildomatic_appserverdir}/webapps/jasperserver) {
+    file { 'default_master.properties':
+      ensure  => present,
+      path    => "/tmp/jasperreports-server-cp-${pkg_version}-bin/buildomatic/default_master.properties",
+      owner   => $buildomatic_user,
+      group   => $buildomatic_user,
+      content => template('jasperreports_server/default_master.properties.erb'),
+      require => [ Archive::Nexus["/tmp/jasperreports-server-cp-${pkg_version}-bin.zip"], ],
+    } ->
+    # Run the js-install with minimal flag
+    exec { 'Run js-install minimal':
+      path    => "/bin:/usr/bin:/sbin:/usr/sbin:/tmp/jasperreports-server-cp-${pkg_version}-bin/buildomatic",
+      cwd     => "/tmp/jasperreports-server-cp-${pkg_version}-bin/buildomatic",
+      command => 'js-install-ce.sh minimal',
+      creates => "${buildomatic_appserverdir}/webapps/jasperserver",
+      user    => $buildomatic_user,
+    } ->
+    # Dirty hack because of issues getting js-install to run as non-root user
+    exec { 'Update Jasper WebApp Ownership':
+      path        => '/bin:/usr/bin:/sbin:/usr/sbin',
+      command     => "chown -R tomcat:tomcat ${buildomatic_appserverdir}/webapps/jasperserver",
+      unless      => "stat -c %U ${buildomatic_appserverdir}/webapps/jasperserver | grep -q 'tomcat'",
+      subscribe   => Exec['Run js-install minimal'],
+      refreshonly => true,
+    }
   }
-
 }
